Do the DB deletion in post-teardown. Tests run without babysitting now but are still noisy.
Post-teardown is when the superclass closes its connections, which inadvertantly re-creates the SQLite files if we had previously tried to delete them (like in a plain old tearDown method).

Also, realize the MisconfiguredTest isn't a proper subclass of PinDbTestCase, which simplifies the implementation. This makes the _fixture_setup and _fixture_teardown overrides go away.